stock-prediction-mlops/requirements.txt
pandas
numpy
scikit-learn
tsfresh
pandas-datareader
boto3
pyyaml
fastapi
uvicorn
mlflow
evidently
lakefs
joblib
lakefs-client